% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregated_ageg_name.R
\name{aggregated_ageg_name}
\alias{aggregated_ageg_name}
\title{Compute the aggregated age group names from a vector of
more detailed age groups}
\usage{
aggregated_ageg_name(selected.ageg, ageg.sep = "_")
}
\arguments{
\item{selected.ageg}{character, the fine grain age group vector}

\item{ageg.sep}{character, the ageg group separator character}
}
\value{
character, the name of the aggregated age group
}
\description{
Compute the aggregated age group names from a vector of
more detailed age groups
}
\examples{
ageg.in <- c('15_19', '20_24', '25_29')
aggregated_ageg_name(ageg.in)

}
