% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop_vector.R
\name{chop_vector}
\alias{chop_vector}
\title{Generate all the possible combinations of slices in a chopped vector}
\usage{
chop_vector(vect.size = 3)
}
\arguments{
\item{vect.size}{int, the size of the vector}
}
\value{
a matrix containing all the possible slices to chope a vector per line
}
\description{
Generate all the possible combinations of slices in a chopped vector
}
\examples{
chop_vector(3)

}
