% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chopped.vector.list}
\alias{chopped.vector.list}
\title{Pre-computed chopped combination for vectors size 1 to 18}
\format{
A 18 item list:
\describe{
each element is a matrix containing all the possible chop combinations
to aggregate a vector of size n.
...
}
}
\usage{
chopped.vector.list
}
\description{
This is a list containing all the possible combination of slices to chop
vectors of size 1 to 18. It is useful to compute custom age group
aggregation to ensure we are not disclosing age group with too few cancer
cases.
}
\concept{data}
\keyword{datasets}
