% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_registry_input_file.R
\name{create_registry_input_file}
\alias{create_registry_input_file}
\title{Create a template file to be filled by cancer registry partners}
\usage{
create_registry_input_file(filename = "cancerRADAR_input.xlsx", verbose = TRUE)
}
\arguments{
\item{filename}{file path, the name of the template file to be created}

\item{verbose}{logical, shall progress message be printed}
}
\value{
a template .xlsx file is created on the hard drive.
}
\description{
Create a template file to be filled by cancer registry partners
}
\examples{
\donttest{
  file.in <- 'input_file_test.xlsx'
  create_registry_input_file(file.in)
  ## remove the file to pass package computation tests
  unlink(file.in)
}
}
\concept{main}
