% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_static_report.R
\name{create_static_report}
\alias{create_static_report}
\title{Create a static report from cancer RADAR output file}
\usage{
create_static_report(filename.out = "")
}
\arguments{
\item{filename.out}{file path, the path to a cancer RADAR output file

This function will create a html report that could be useful to check the data
that will be transmitted to IARC.}
}
\value{
nothing is returned, but a html file created with some summary statistics and graphs
out of the file that should be transmitted with IARC
}
\description{
Create a static report from cancer RADAR output file
}
