% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.aggr}
\alias{dat.aggr}
\title{Geographical aggregation used for cancerradarr}
\format{
A data frame with 250 rows and 5 columns:
\describe{
\item{cob_iso3}{Country ISO3 code}
\item{un_region}{UN region}
\item{un_subregion}{UN subregion}
\item{hdi_cat}{HDI 2023 category}
\item{any_migr}{any migration background}
...
}
}
\usage{
dat.aggr
}
\description{
In order to prevent loose of data in case of too low effective, several geographical aggregation
can be considered. In this table are stored the different level of aggregation and the aggregation
correspondence table considered.
}
\details{
A multi-columns dataset containing all the countries of birth (as ISO3 code) and other geographical
aggregation rules
}
\concept{data}
\keyword{datasets}
