% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.cob}
\alias{dat.cob}
\title{Countries label and countries codes}
\format{
A data frame with 251 rows and 3 columns:
\describe{
\item{cob_label}{Country name}
\item{cob_code}{Country code}
\item{cob_iso3}{Country ISO3 code (used as unique id)}
...
}
}
\usage{
dat.cob
}
\description{
A 2 column dataset containing all the countries of birth (with associated countries codes) included
in Cancer RADAR project
}
\concept{data}
\keyword{datasets}
