% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_canradar_dictionary.R
\name{open_canradar_dictionary}
\alias{open_canradar_dictionary}
\title{Open cancer RADAR output file dictionary}
\usage{
open_canradar_dictionary()
}
\value{
the path to a temporary file where cancer RADAR dictionary is stored
}
\description{
Calling this function will open the dictionary describing sheets and variables stored in the cancer
summary file (output file generated by \code{\link{create_canradar_summary_file}} function). It
could be useful for the cancer registries to check what kind of data they will be sharing.
Note that a temporary copy of the dictionary is created on your hard drive to prevent from
unwanted file modification.
}
\examples{
\dontrun{
   open_canradar_dictionary()
}
}
