% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dynamic_report.R
\name{run_dynamic_report}
\alias{run_dynamic_report}
\title{Create a dynamic report from cancer RADAR output file}
\usage{
run_dynamic_report(filename.out = "")
}
\arguments{
\item{filename.out}{file path, the path to a cancer RADAR output file

This function will open a shiny app where cancer registries can visually check the data
they will be transmitted to IARC.}
}
\value{
nothing is returned
}
\description{
Create a dynamic report from cancer RADAR output file
}
\concept{main}
